Change Aux BMS charge allowance to comply with scrutineering
According to the VDR - there are 3 events in which the BPS prevents:

Over Current, Under Voltage & Over Voltage should turn off all the contactors.

Over Temperature: Our max charging temp is 45, and the max discharge is 60. If we are out of out charging range but before discharge, it should just disengage the charge contactor.

If the charge contactor is opened due to a temperature trip BUT there is still a negative current applied (IE Car is still somehow charging), the car should trip.

Over/Under Voltage: The car should open the contactors when we hit our upper/lower internal limits(> 4.15 & < 2.60), but trip when the Orion wants it to turn off - above 4.2 and below 2.55.

Over Current: The car should trip when we hit a current below -47 A (charge current) or above 229 A (dischage current).

Also, we dropped functionality for differentiating driverInputs.